+++ js/main.js
@@ -1,5 +1,4 @@
-import renderScreen from './render-screen.js';
-import startScreen from './screen-intro.js';
-import {INITIAL_GAME} from './data/game-mechanics.js';
+import show from './render-screen.js';
+import screenIntro from './screen-intro.js';
 
-renderScreen(startScreen(INITIAL_GAME));
+show(screenIntro().element);
