+++ js/screen-intro-view.js
@@ -1,0 +1,29 @@
+import AbstractView from './abstract-view.js';
+
+const nextButtonClass = `.intro__asterisk`;
+
+export default class ScreenIntroView extends AbstractView {
+  constructor() {
+    super();
+  }
+
+  get template() {
+    return `
+    <section class="intro">
+      <button class="intro__asterisk asterisk" type="button"><span class="visually-hidden">Продолжить</span>*</button>
+      <p class="intro__motto"><sup>*</sup> Это не фото. Это рисунок маслом нидерландского художника-фотореалиста Tjalf Sparnaay.</p>
+    </section>
+    `;
+  }
+
+  bind(screen) {
+    screen.querySelector(nextButtonClass).addEventListener(`click`, (evt) => {
+      evt.stopPropagation();
+      evt.preventDefault();
+      this.onNext();
+    });
+  }
+
+  onNext() {
+  }
+}
