+++ js/screen-rules.js
@@ -1,43 +1,14 @@
-import {createMarkupNode} from './utils.js';
 import header from './screen-header.js';
-import {startGame} from './screen-game.js';
+import screenGame from './screen-game.js';
+import show from './render-screen.js';
+import {INITIAL_GAME} from './data/game-mechanics.js';
+import ScreenRulesView from './screen-rules-view.js';
 
-export default (state) => {
-  const template = `
-    <section class="rules">
-      <h2 class="rules__title">Правила</h2>
-      <ul class="rules__description">
-        <li>Угадай 10 раз для каждого изображения фото
-          <img class="rules__icon" src="img/icon-photo.png" width="32" height="31" alt="Фото"> или рисунок
-          <img class="rules__icon" src="img/icon-paint.png" width="32" height="31" alt="Рисунок"></li>
-        <li>Фотографиями или рисунками могут быть оба изображения.</li>
-        <li>На каждую попытку отводится 30 секунд.</li>
-        <li>Ошибиться можно не более 3 раз.</li>
-      </ul>
-      <p class="rules__ready">Готовы?</p>
-      <form class="rules__form">
-        <input class="rules__input" type="text" placeholder="Ваше Имя">
-        <button class="rules__button  continue" type="submit" disabled>Go!</button>
-      </form>
-    </section>
-  `;
+export default () => {
+  const node = new ScreenRulesView();
+  node.element.insertAdjacentElement(`afterbegin`, header().element);
 
-  const node = createMarkupNode(template);
-  node.insertAdjacentElement(`afterbegin`, header(state));
-
-  const nameInputField = node.querySelector(`.rules__input`);
-  const userDataForm = node.querySelector(`.rules__form`);
-  const gameStartButton = node.querySelector(`.rules__button`);
-
-  userDataForm.addEventListener(`submit`, (evt) => {
-    evt.preventDefault();
-
-    startGame();
-  });
-
-  nameInputField.addEventListener(`input`, () => {
-    gameStartButton.disabled = !nameInputField.value;
-  });
+  node.onNext = () => show(screenGame(Object.assign({}, INITIAL_GAME, {answers: []})).element);
 
   return node;
 };
